Moving set/removeEventProperty to PropertyConfigurator
