+++ sdk/appcenter-analytics/src/test/java/com/microsoft/appcenter/analytics/AnalyticsTransmissionTargetTest.java
@@ -8,15 +8,11 @@ import com.microsoft.appcenter.ingestion.models.Log;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.mockito.ArgumentCaptor;
 import org.mockito.ArgumentMatcher;
 import org.mockito.Mock;
 
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.Map;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
@@ -287,152 +283,5 @@ public class AnalyticsTransmissionTargetTest extends AbstractAnalyticsTest {
         assertFalse(grandParentTarget.isEnabledAsync().get());
         assertFalse(parentTarget.isEnabledAsync().get());
         assertFalse(childTarget.isEnabledAsync().get());
-    }
-
-    @Test
-    public void setCommonEventProperties() {
-
-        /* Create transmission target and add property. */
-        AnalyticsTransmissionTarget target = Analytics.getTransmissionTarget("test");
-        target.setEventProperty("key", "value");
-
-        /* Track event without property. */
-        target.trackEvent("eventName");
-        verify(mChannel).enqueue(argThat(new ArgumentMatcher<Log>() {
-
-            @Override
-            public boolean matches(Object item) {
-                if (item instanceof EventLog) {
-                    EventLog eventLog = (EventLog) item;
-                    Map<String, String> expectedProperties = new HashMap<>();
-                    expectedProperties.put("key", "value");
-                    boolean nameAndPropertiesMatch = eventLog.getName().equals("eventName") && expectedProperties.equals(eventLog.getProperties());
-                    boolean tokenMatch = eventLog.getTransmissionTargetTokens().size() == 1 && eventLog.getTransmissionTargetTokens().contains("test");
-                    return nameAndPropertiesMatch && tokenMatch;
-                }
-                return false;
-            }
-        }), anyString());
-    }
-
-    @Test
-    public void setAndRemoveCommonEventPropertiesWithMerge() {
-
-        /* Create transmission target and add 2 properties (1 overwritten). */
-        AnalyticsTransmissionTarget target = Analytics.getTransmissionTarget("test");
-        target.setEventProperty("key1", "value1");
-        target.setEventProperty("key2", "ignore");
-        target.setEventProperty("remove", "ignore");
-
-        /* Remove some properties. */
-        target.removeEventProperty("remove");
-        target.removeEventProperty("notFound");
-
-        /* Prepare properties. */
-        Map<String, String> properties = new HashMap<>();
-        properties.put("key2", "value2");
-        properties.put("key3", "value3");
-
-        /* Track event with extra properties. */
-        target.trackEvent("eventName", properties);
-        verify(mChannel).enqueue(argThat(new ArgumentMatcher<Log>() {
-
-            @Override
-            public boolean matches(Object item) {
-                if (item instanceof EventLog) {
-                    EventLog eventLog = (EventLog) item;
-                    Map<String, String> expectedProperties = new HashMap<>();
-                    expectedProperties.put("key1", "value1");
-                    expectedProperties.put("key2", "value2");
-                    expectedProperties.put("key3", "value3");
-                    boolean nameAndPropertiesMatch = eventLog.getName().equals("eventName") && expectedProperties.equals(eventLog.getProperties());
-                    boolean tokenMatch = eventLog.getTransmissionTargetTokens().size() == 1 && eventLog.getTransmissionTargetTokens().contains("test");
-                    return nameAndPropertiesMatch && tokenMatch;
-                }
-                return false;
-            }
-        }), anyString());
-    }
-
-    @Test
-    public void trackEventWithEmptyProperties() {
-
-        /* Create transmission target. */
-        AnalyticsTransmissionTarget target = Analytics.getTransmissionTarget("test");
-
-        /* Track event with empty properties. */
-        target.trackEvent("eventName", Collections.<String, String>emptyMap());
-        verify(mChannel).enqueue(argThat(new ArgumentMatcher<Log>() {
-
-            @Override
-            public boolean matches(Object item) {
-                if (item instanceof EventLog) {
-                    EventLog eventLog = (EventLog) item;
-                    Map<String, String> expectedProperties = Collections.emptyMap();
-                    boolean nameAndPropertiesMatch = eventLog.getName().equals("eventName") && expectedProperties.equals(eventLog.getProperties());
-                    boolean tokenMatch = eventLog.getTransmissionTargetTokens().size() == 1 && eventLog.getTransmissionTargetTokens().contains("test");
-                    return nameAndPropertiesMatch && tokenMatch;
-                }
-                return false;
-            }
-        }), anyString());
-    }
-
-    @Test
-    public void eventPropertiesCascading() {
-
-        /* Create transmission target hierarchy. */
-        AnalyticsTransmissionTarget grandParent = Analytics.getTransmissionTarget("grandParent");
-        AnalyticsTransmissionTarget parent = grandParent.getTransmissionTarget("parent");
-        AnalyticsTransmissionTarget child = parent.getTransmissionTarget("child");
-
-        /* Set common properties across hierarchy with some overrides. */
-        grandParent.setEventProperty("a", "1");
-        grandParent.setEventProperty("b", "2");
-        grandParent.setEventProperty("c", "3");
-
-        /* Override some. */
-        parent.setEventProperty("a", "11");
-        parent.setEventProperty("b", "22");
-
-        /* And a new one. */
-        parent.setEventProperty("d", "44");
-
-        /* Just to show we still get value from grandParent if we remove an override. */
-        parent.setEventProperty("c", "33");
-        parent.removeEventProperty("c");
-
-        /* Overrides in child. */
-        child.setEventProperty("d", "444");
-
-        /* New in child. */
-        child.setEventProperty("e", "555");
-        child.setEventProperty("f", "666");
-
-        /* Track event in child. Override properties in trackEvent. */
-        Map<String, String> properties = new HashMap<>();
-        properties.put("f", "6666");
-        properties.put("g", "7777");
-        child.trackEvent("eventName", properties);
-
-        /* Verify log that was sent. */
-        ArgumentCaptor<EventLog> logArgumentCaptor = ArgumentCaptor.forClass(EventLog.class);
-        verify(mChannel).enqueue(logArgumentCaptor.capture(), anyString());
-        EventLog log = logArgumentCaptor.getValue();
-        assertNotNull(log);
-        assertEquals("eventName", log.getName());
-        assertEquals(1, log.getTransmissionTargetTokens().size());
-        assertTrue(log.getTransmissionTargetTokens().contains("child"));
-
-        /* Verify properties. */
-        Map<String, String> expectedProperties = new HashMap<>();
-        expectedProperties.put("a", "11");
-        expectedProperties.put("b", "22");
-        expectedProperties.put("c", "3");
-        expectedProperties.put("d", "444");
-        expectedProperties.put("e", "555");
-        expectedProperties.put("f", "6666");
-        expectedProperties.put("g", "7777");
-        assertEquals(expectedProperties, log.getProperties());
     }
 }
